fix:sonarqube_security_hotspot
