+++ frontend/src/components/forms/shared/formValidationUtils.ts
@@ -1,3 +1,4 @@
+import validator from 'validator'
 import type { ValidationRule } from 'components/forms/shared/useFormValidation'
 
 export const validateRequired = (value: string, fieldName: string): string | undefined => {
@@ -30,6 +31,35 @@ export const validateEndDate = (value: string, startDate?: string): string | und
     return 'End date must be after start date'
   }
   return undefined
+}
+
+export const validateWebsite = (value: string): string | undefined => {
+  const requiredError = validateRequired(value, 'Website')
+  if (requiredError) return requiredError
+
+  try {
+    const parsedUrl = new URL(value)
+    if (!['http:', 'https:'].includes(parsedUrl.protocol)) {
+      return 'Website must start with http:// or https://'
+    }
+  } catch {
+    return 'Enter a valid website URL'
+  }
+
+  return undefined
+}
+
+export const validateContactEmail = (value: string): string | undefined => {
+  const requiredError = validateRequired(value, 'Contact email')
+  if (requiredError) return requiredError
+  if (!validator.isEmail(value.trim())) {
+    return 'Enter a valid contact email'
+  }
+  return undefined
+}
+
+export const validateMessage = (value: string): string | undefined => {
+  return validateRequired(value, 'Message')
 }
 
 type CommonFormData = {
